moved the charts folder to the coinsPage folder, also renamed the components
