--- src/components/coinsPage/CarouselSlider/CarouselSliderServer.tsx
@@ -1,10 +1,0 @@
-import { getCarouselSliderData } from "@/actions/getCoinMarketData";
-import CarouselSlider from ".";
-
-const CarouselSliderServer = async () => {
-  const data = await getCarouselSliderData();
-
-  return <CarouselSlider data={data} />;
-};
-
-export default CarouselSliderServer;
