+++ src/components/coinsPage/CoinSlider/CarouselTitle.tsx
