+++ src/components/coinsPage/ComparisonCharts/AreaChart/index.tsx
