+++ src/components/coinsPage/ComparisonCharts/BarChart/index.tsx
