+++ src/components/coinsPage/ComparisonCharts/IntervalTabs.tsx
